Create new folder src  and move App.tsx file to src
For linting purpose
